fix warnings

+++ Mesh_3/include/CGAL/exude_mesh_3.h
@@ -96,7 +96,7 @@ Mesh_optimization_return_code exude_mesh_3(C3T3& c3t3,const CGAL_NP_CLASS& np = 
 {
   using parameters::choose_parameter;
   using parameters::get_parameter;
-  int time_limit=choose_parameter(get_parameter(np,internal_np::maximum_running_time),parameters::default_values_for_mesh_3::time_limit);
+  double time_limit=choose_parameter(get_parameter(np,internal_np::maximum_running_time),parameters::default_values_for_mesh_3::time_limit);
   double sliver_bound= choose_parameter(get_parameter(np,internal_np::lower_sliver_bound),parameters::default_values_for_mesh_3::exude_sliver_bound);
   return exude_mesh_3_impl(c3t3,time_limit,sliver_bound);
 }
